drm/bridge: synopsys: dw-hdmi: Fix hdmi color can't be switched when entering kernel
In scenarios where U-boot can read EDID but the kernel cannot,
HDMI color of U-boot and the kernel may differ. The current color
switching process causes this scenario where HDMI enters the
kernel but fails to switch to the correct color. Optimize process
of switching colors to solve this problem.

Change-Id: I36c7d7e68e438bd33f9d2804d4a5a34edda9ba4a
Signed-off-by: Algea Cao <algea.cao@rock-chips.com>