drm/rockchip: dw-dp: add a debugfs directory
The DPTX may be not create a drm connector, or it will create
more than 1 drm connector when work in MST mode. So it's better
to create a debugfs directory for the DPTX driver. A drm
encoder is necessary for DPTX driver and the drm connector is
optional, So the debug node add when drm encoder late register
state.

Change-Id: I7515872aebc0101611a9afa393765a8bd478699d
Signed-off-by: Zhang Yubing <yubing.zhang@rock-chips.com>